+++ src/main/java/com/nisovin/shopkeepers/commands/shopkeepers/CommandGive.java
@@ -1,5 +1,7 @@
 package com.nisovin.shopkeepers.commands.shopkeepers;
 
+import java.util.Arrays;
+
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
@@ -15,7 +17,10 @@ import com.nisovin.shopkeepers.commands.lib.arguments.DefaultValueFallback;
 import com.nisovin.shopkeepers.commands.lib.arguments.PlayerArgument;
 import com.nisovin.shopkeepers.commands.lib.arguments.PositiveIntegerArgument;
 import com.nisovin.shopkeepers.commands.lib.arguments.SenderPlayerFallback;
+import com.nisovin.shopkeepers.spigot.text.SpigotText;
+import com.nisovin.shopkeepers.spigot.text.Text;
 import com.nisovin.shopkeepers.util.ItemUtils;
+import com.nisovin.shopkeepers.util.MapUtils;
 import com.nisovin.shopkeepers.util.TextUtils;
 
 class CommandGive extends Command {
@@ -62,8 +67,10 @@ class CommandGive extends Command {
 		}
 
 		// success:
-		TextUtils.sendMessage(sender, Settings.msgShopCreationItemsGiven,
-				"{player}", targetPlayer.getName(),
-				"{amount}", String.valueOf(amount));
+		// TODO show shop creation item via hover text?
+		SpigotText.sendMessage(sender, Settings.msgShopCreationItemsGiven, Arrays.asList(
+				MapUtils.entry("{player}", TextUtils.getPlayerText(targetPlayer)),
+				MapUtils.entry("{amount}", new Text(String.valueOf(amount)))
+		));
 	}
 }