+++ src/main/java/com/nisovin/shopkeepers/villagers/VillagerInteractionListener.java
@@ -91,6 +91,7 @@ public class VillagerInteractionListener implements Listener {
 		PlayerInventory playerInventory = player.getInventory();
 		ItemStack itemInMainHand = playerInventory.getItemInMainHand();
 		if (!Settings.isHireItem(itemInMainHand)) {
+			// TODO show hire item via hover event?
 			TextUtils.sendMessage(player, Settings.msgVillagerForHire, "{costs}", String.valueOf(Settings.hireOtherVillagersCosts),
 					"{hire-item}", Settings.hireItem.getType().name()); // TODO also print required hire item name and lore?
 			return false;